refactor: split claims validate-quality.ts into 4 focused modules
Split the 725-line monolithic validate-quality.ts into a directory with
4 focused modules, following the TODO already in the file:

- types.ts (98 lines): CheckName, CheckResult, interfaces, patterns
- checks.ts (378 lines): 10 check functions + helper utilities
- report.ts (85 lines): human and JSON report formatters
- index.ts (207 lines): orchestrator, main entry point, re-exports

No logic changes. All code moved verbatim with only import paths
adjusted. The script entry point in commands/claims.ts updated to
point to the new index.ts.

Closes #1287
